requirements.txt
asgiref==3.7.2
Django==4.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.2
typing_extensions==4.9.0
uritemplate==4.1.1
